2007-12-14  Sancho Lerena <slerena@gmail.com>
        * pandoradb.sql: BUG FIXED, generic_data_inc data don't work with 
        small float data. Need to alter tagente_datos_inc table for field
        "datos.



git-svn-id: https://svn.code.sf.net/p/pandora/code/trunk@681 c3f86ba8-e40f-0410-aaad-9ba5e7f4b01f

+++ pandora_console/pandoradb.sql
@@ -62,7 +62,7 @@ CREATE TABLE `tagente_datos` (
 CREATE TABLE `tagente_datos_inc` (
   `id_adi` bigint(20) unsigned NOT NULL auto_increment,
   `id_agente_modulo` bigint(20) NOT NULL default '0',
-  `datos` bigint(12) default NULL,
+  `datos` double(18,2) default NULL,
   `timestamp` datetime NOT NULL default '0000-00-00 00:00:00',
   `utimestamp` int(10) unsigned default '0',
   PRIMARY KEY  (`id_adi`),
